Have the dell n3000 and dell powerconnect switches use different extensions in the config file
* Earlier, both dell switches were sharing a single extension setting
(hil.ext.switches.dell). This commit separates that.

* Added a helper method `_disconnect()` that takes care of disconnecting in
console based switches.

* Move disconnect() into dell.py and n3000.py, and in that the drivers call
_should_save with the right argument for that switch before calling _disconnect

* Update the saving-to-switch-flash deployment tests to include n3000 switch.
